Refactor get_oneoffixx_templates to make it reusable
